test(serde_yml): add test for more complex nested merges
This test ensures that nested merges are handled in a depth-first manner,
that resolves deeper nesting before resolving shallower nesting.
Right now this test fails like this:

  assertion `left == right` failed
    left: Mapping {"a": Number(1), "b": Number(2)}
   right: Mapping {"<<": Mapping {"a": Number(1)}}

because the "<<" key from `nest_a` is merged into `outer` as-is
(even though it should instead be merged into `outer`)
and then the "<<" key from `nest_b` is _not_ merged because the
key is already present in `outer`.

This means the fields from `nest_b` are now lost and the correct merge
can not be achieved by just resolving the "<<" in `outer` after the fact.

Instead the merges coming from `nest_a` and `nest_b` have to be resolved
first and only then can the `outer` merge be applied.